se valida si existe el archivo para eliminar la foto de perfil

+++ app/Http/Controllers/ApiAuthController.php
@@ -239,13 +239,14 @@ class ApiAuthController extends Controller
             $image->move($destinationPath, $input['imagename']);
 
             $account = Account::byCI($request->ci);
-            if($account->photo == 1)
+
+            $filename = $destinationPath.$account->namePhoto;
+
+            //verifica si ya tiene una foto guardada el usuario y la borra
+            if($account->photo == 1 && File::exists($filename))
             {
-                //verifica si ya tiene una foto guardada el usuario y la borra
-                $filename = $destinationPath.$account->namePhoto;
                 File::delete($filename);
                 //unlink($filename);
-
             }else{
                 $account->photo = 1;
             }
